+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to Starchild Music will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.3] - 2025-12-09
+
+### Changed
+
+- Version bump to 0.5.3
+
 ## [0.5.2] - 2025-12-09
 
 ### Added
@@ -19,9 +25,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.5.1] - 2025-12-09
 
-### Added
+### Changed
 
-#### Logging Enhancements
+- Voronoi pattern smoothing & soothing
+- Removed Plasma pattern completely, it was out of place, too in your face
+- Bubble pattern replaced with The Orb
+- Flow field pattern optimized for performance
+- Reduced Mandala fidelity, layering and symmetry. Sometimes less is more.
+
+#### Logging Enhancements & Visualizer Pattern Changes
 
 - **Visual Component Logging**: Added comprehensive logging for visual components and visualizer patterns
   - Logging for visual pattern transitions and rendering